refactor: 관련 파일끼리 폴더 구성
